+++ .gitignore
@@ -10,6 +10,6 @@ data/
 
 # datafusion
 .history
+
+# macOS
 .DS_Store
-optd/.DS_Store
-optd/statistics/.DS_Store
