+++ optd/catalog/Cargo.toml
@@ -5,3 +5,10 @@ edition.workspace = true
 repository.workspace = true
 
 [dependencies]
+serde = { version = "1.0", features = ["derive"] }
+duckdb = { version = "1.4.0", features = ["bundled"] }
+snafu = "0.8.6"
+serde_json = "1.0"
+
+[dev-dependencies]
+tempfile = "3.8"
