--- optd/statistics/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "optd-statistics"
-version.workspace = true
-edition.workspace = true
-repository.workspace = true
-
-[dependencies]
-parking_lot = "=0.12.5"
-glob = "0.3.2"
-url = { version = "^2.5.7" }
-async-trait = { version = "^0.1.89" }
-tokio = { workspace = true, features = ["full"] }
-duckdb = { features = ["bundled"], version = "1.4.0" }
-snafu = "0.8.6"
-serde_json = "1.0"
-futures = "0.3.31"
-serde = "1.0.0"
-
-[dev-dependencies]
-tempfile = "3.8"
