--- optd/statistics/src/lib.rs
@@ -1,3 +1,0 @@
-mod statistics;
-
-pub use statistics::{DuckLakeStatisticsProvider, Error as InterfaceError, StatisticsProvider};
