--- optd/storage/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "optd-storage"
-version.workspace = true
-edition.workspace = true
-repository.workspace = true
-
-[dependencies]
